move code from mkmf.rb to rbconfig.rb, where it belongs
